+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "http": "^0.0.1-security"
+  }
+}
